Code at end of backend-chapter-6

+++ MagicMirror.DataAccess.Entities/Weather/Main.cs
@@ -2,7 +2,7 @@
 {
     public class Main
     {
-        public float Temp { get; set; }
+        public double Temp { get; set; }
         public int Pressure { get; set; }
         public int Humidity { get; set; }
         public float Temp_min { get; set; }
